third_party/blink/web_tests/external/wpt/WebIDL/ecmascript-binding/class-string-iterator-prototype-object.any.worker-expected.txt

This is a testharness.js-based test.
FAIL @@toStringTag exists with the appropriate descriptor assert_equals: value expected "URLSearchParams Iterator" but got "Iterator"
FAIL Object.prototype.toString assert_equals: expected "[object URLSearchParams Iterator]" but got "[object Iterator]"
PASS Object.prototype.toString applied after modifying @@toStringTag
FAIL Object.prototype.toString applied after nulling the prototype assert_equals: expected "[object Object]" but got "[object URLSearchParams Iterator]"
PASS Object.prototype.toString applied after deleting @@toStringTag
Harness: the test ran to completion.
